rootdev: support btrfs
With btrfs it's not possible in general to find a single underlying
block device because there could be many, but still it is possible for
our case since we have a single partition. However, rootdev failed to
find it because it uses the device minor and major pair from a "stat"
syscall on the mount point. This doesn't work with btrfs because it
gives a logical pair instead of one belonging to a block device.

Instead of using the minor and major pair, find the underlying block
device of a mount point by looking it up in /proc/mounts. The existing
traversal code to find the lowest backing device is kept in place by
getting the minor and major pair for the found block device.