rootdev: Only use stat.rdev instead of looking up /sys/block/DEV/dev
When /dev/mapper/usr is not a symlink to /dev/dm-0 but directly an equal
device node, we can't find the /sys/block/DEV/dev entry for it because
this is /sys/block/dm-0/dev and we can't device the "dm-0" name.
However, we can directly use the device major and minor numbers of the
device node through stat and continue with it in any case.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>